requirements.txt
streamlit==1.30.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.10
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
chromadb==0.4.22
langchain-text-splitters>=0.0.1
pydantic>=2.0.0
urllib3==2.1.0
lxml==4.9.3